refactor: split into test and non-test file
